Changed ints to doubles and reflecting ES while NP is rotated

+++ src/main/java/domain/ability/HollowPurple.java
@@ -6,7 +6,7 @@ import domain.needForSpear.Statistics;
 import java.util.Random;
 
 public class HollowPurple extends Ability {
-    public final int gameScreenWidth = Controller.getInstance().getFrameBorders()[0],
+    public final double gameScreenWidth = Controller.getInstance().getFrameBorders()[0],
             gameScreenHeight = Controller.getInstance().getFrameBorders()[1],
             hollowPurpleNum = 8;
     public Random randi = new Random();
@@ -16,7 +16,7 @@ public class HollowPurple extends Ability {
         super();
         //The GameScreen is divided into cells (size: 10x10) where hollow purple obstacles can be put. The objects are
         // put into the empty cells chosen randomly in a way that they won't clash with already existing obstacles.
-        boolean[][] locationCells = new boolean[(gameScreenHeight-200)/40][gameScreenWidth/100];
+        boolean[][] locationCells = new boolean[(int) (gameScreenHeight-200)/40][(int)gameScreenWidth/100];
         for(int i=0; i<hollowPurpleNum; i++){
             putHPInCell(locationCells);
         }
