+++ src/main/java/domain/body/Body.java
@@ -1,35 +1,35 @@
 package domain.body;
 
 public class Body {
-    protected int x;
-    protected int y;
-    public int width;
-    public int height;
+    protected double x;
+    protected double y;
+    public double  width;
+    public double  height;
 
-    public Body(int x_coordinates,
-                int y_coordinates,
-                int width,
-                int height) {
+    public Body(double x_coordinates,
+                double y_coordinates,
+                double width,
+                double height) {
         x = x_coordinates;
         y = y_coordinates;
         this.width = width;
         this.height = height;
     }
     
-    public int[] getCoordinates(){
-        return new int[] {x,y};
+    public double[] getCoordinates(){
+        return new double[] {x,y};
     }
-    public void setCoordinates(int x, int y) {
+    public void setCoordinates(double  x, double  y) {
         this.x = x;
         this.y = y;
     }
-    public boolean compareCoordinates(int x, int y, int width, int height) {
+    public boolean compareCoordinates(double x, double  y, double width, double height) {
         boolean crashingCoordinates = false;
-        int x1 = this.x;
-        int y1 = this.y;
+        double x1 = this.x;
+        double y1 = this.y;
 
-        int height1 = this.height;
-        int width1 = this.width;
+        double height1 = this.height;
+        double width1 = this.width;
 
         if(x >= x1 && x-x1 <= width1 && y >= y1 && y-y1 <= height1){
             crashingCoordinates = true;
