+++ src/main/java/domain/body/obstacle/FirmObstacle.java
@@ -9,10 +9,10 @@ public class FirmObstacle extends Obstacle {
 
     public boolean movesRight;
 
-    public FirmObstacle(int x_coordinates,
-                        int y_coordinates,
-                        int width,
-                        int height,
+    public FirmObstacle(double x_coordinates,
+                        double y_coordinates,
+                        double width,
+                        double height,
                         int numberOfHits) {
         super(x_coordinates, y_coordinates, width, height, numberOfHits);
         //Speed is initialized as this because otherwise it is smaller than 1 and doesn't have an effect.
